projects/alpaca-trading/pyproject.toml
[project]
name = "alpaca-trader"
version = "0.1.0"
description = "Automated trading system for Alpaca"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.30.0",
    "pandas>=2.0.0",
    "apscheduler>=3.10.0",
    "ta>=0.11.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
